Slett duplikat MottaSøknadController etter master merge
